# Optimize away duplicate address loading instructions.