--- a/Backend/Engine/NNUE/Architecture/Basic/BasicNNUE.cs
+++ b/Backend/Engine/NNUE/Architecture/Basic/BasicNNUE.cs
@@ -39,9 +39,6 @@ public class BasicNNUE
     private readonly short[] BlackPOV = new short[INPUT];
 
     private readonly BasicAccumulator<short>[] Accumulators = new BasicAccumulator<short>[ACCUMULATOR_STACK_SIZE];
-    
-    // ReSharper disable once UnusedMember.Local
-    private readonly short[] Flatten = new short[HIDDEN * 2];
 
     private readonly int[] Output = new int[OUTPUT];
     
@@ -162,19 +159,8 @@ public class BasicNNUE
     [MethodImpl(MethodImplOptions.AggressiveOptimization)]
     public int Evaluate(PieceColor colorToMove)
     {
-#if DEBUG
-        int firstOffset = 0;
-        int secondOffset = 256;
-
-        if (colorToMove == PieceColor.Black) {
-            firstOffset = 256;
-            secondOffset = 0;
-        }
-#endif
-        
         BasicAccumulator<short> accumulator = Accumulators.AA(CurrentAccumulator);
 
-#if RELEASE
         if (colorToMove == PieceColor.White) {
             NN.ClippedReLUFlattenAndForward(accumulator.A, accumulator.B, FeatureBias, OutWeight, Output, 
                 CR_MIN, CR_MAX, HIDDEN);
@@ -182,14 +168,6 @@ public class BasicNNUE
             NN.ClippedReLUFlattenAndForward(accumulator.B, accumulator.A, FeatureBias, OutWeight, Output, 
                 CR_MIN, CR_MAX, HIDDEN);
         }
-#endif
-        
-#if DEBUG
-        NN.ClippedReLU(accumulator.A, FeatureBias, Flatten, CR_MIN, CR_MAX, firstOffset);
-        NN.ClippedReLU(accumulator.B, FeatureBias, Flatten, CR_MIN, CR_MAX, secondOffset);
-        
-        NN.Forward(Flatten, OutWeight, Output);
-#endif
         
         return (Output.AA(0) + OutBias.AA(0)) * SCALE / QAB;
     }