--- a/Backend/Engine/NNUE/Vectorization/VMethod.cs
+++ b/Backend/Engine/NNUE/Vectorization/VMethod.cs
@@ -16,9 +16,21 @@ public static class VMethod
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static Vector<T> LoadVector<T>(this ref T value, int index = 0) where T : struct
+    {
+        return Unsafe.ReadUnaligned<Vector<T>>(ref Unsafe.As<T, byte>(ref Unsafe.Add(ref value, index)));
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static void ToArray<T>(this Vector<T> vector, T[] array, int offset = 0) where T : struct
     {
         Unsafe.WriteUnaligned(ref Unsafe.As<T, byte>(ref array.AA(offset)), vector);
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static void ToArray<T>(this Vector<T> vector, ref T arrayRef, int offset = 0) where T : struct
+    {
+        Unsafe.WriteUnaligned(ref Unsafe.As<T, byte>(ref Unsafe.Add(ref arrayRef, offset)), vector);
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]