This PR optimizes away multiple `LEA: Load Effective Address` instructions. This is normally required since .NET GC moves stuff around memory, but the instruction should be only added if it is moved, and not every time.

```
       IN          ACCUMULATOR                               HIDDEN                                    OUT
 ______________      _______      ______________________________________________________________      _____
| WHITE: (768) | -> | (256) | -> | ClippedReLU(0, 1) -> (256) \                                 |    |     |
|              |    |       |    |                             \                                |    |     |
|              |    |       |    |                              CONCATENATE(ColorToMove): (512) | -> | (1) |
|              |    |       |    |                             /                                |    |     |
| BLACK: (768) | -> | (256) | -> | ClippedReLU(0, 1) -> (256) /                                 |    |     |
 --------------      -------      --------------------------------------------------------------      -----

TAG: EXP-0006
Data: 1.7B FEN --- DEPTH: 9 (1.2B) + NODES: 5K (500M)
Batch Size: 500000
Epochs: 80
LR: 5e-3
LR Drop: 30
WDL: 0.4
```

## ELO Difference
Using `UHO_XXL_+0.90_+1.19.epd`:

### TC: 10s + 0.1s (STC)
```
ELO   | 7.02 +- 5.06 (95%)
SPRT  | 10.0+0.10s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 9608 W: 2648 L: 2454 D: 4506
```

### TC: 60s + 0.6s (LTC)
```
ELO   | 5.19 +- 3.41 (95%)
SPRT  | 60.0+0.60s Threads=1 Hash=256MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 3.00]
GAMES | N: 19072 W: 4705 L: 4420 D: 9947
```